Deleted Telop Classes and changed autofunctions update rate to 100 htz (for shooting in teleop)
